MLK-10774-44 imx:mx6 fix is_soc_rev
is_soc_rev should be casted to signed int, otherwise
may incur errors when detecting cpu types.

+++ arch/arm/include/asm/arch-mx6/sys_proto.h
@@ -14,7 +14,7 @@
 #include "../arch-imx/cpu.h"
 
 #define soc_rev() (get_cpu_rev() & 0xFF)
-#define is_soc_rev(rev)        (soc_rev() - rev)
+#define is_soc_rev(rev)        (int)(soc_rev() - rev)
 
 u32 get_nr_cpus(void);
 u32 get_cpu_rev(void);
